backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
beautifulsoup4==4.12.3
lxml==5.3.0
img2pdf==0.5.1
pillow==10.4.0
supabase==2.9.1
python-dotenv==1.0.1
pydantic-settings==2.5.2
playwright==1.50.0
